refactor: Car 클래스 내 String name 변수를 포장하는 Name 클래스 생성
